+++ RAnalogs/examples/rmd-ipy.sh
@@ -8,7 +8,7 @@
 # Weiming: I have set up the environment on Sapphire.
 #
 
-source venv/bin/activate
+# source venv/bin/activate
 
 echo pyrmd --t ipynb --from Rmd -y demo-1_AnEn-basics.Rmd
 ipyrmd --t ipynb --from Rmd -y demo-1_AnEn-basics.Rmd
@@ -22,4 +22,4 @@ ipyrmd --t ipynb --from Rmd -y demo-3_operational-search.Rmd
 echo ipyrmd --t ipynb --from Rmd -y demo-4_SEA2019.Rmd
 ipyrmd --t ipynb --from Rmd -y demo-4_SEA2019.Rmd
 
-deactivate
+# deactivate