+++ NEWS.md
@@ -4,7 +4,8 @@
 - Improve `verifyCRPS`.
 - Improve `RAnEn` function documentation.
 - Add identification for cycle time in C++ utilities.
-- Bug fixed in checking functions.
+- Bug fixed in checking functions in `RAnEn`.
+- Added grid checks in `gribConverter` for each file.
 
 # PAnEn 3.4.2