+++ NEWS.md
@@ -5,7 +5,9 @@
 - Improve `RAnEn` function documentation.
 - Add identification for cycle time in C++ utilities.
 - Bug fixed in checking functions in `RAnEn`.
-- Added grid checks in `gribConverter` for each file.
+- Add grid checks in `gribConverter` for each file.
+- Make `classifier` the parameter for `extendedSearch` only.
+- Improve website and tutorials/posts.
 
 # PAnEn 3.4.2