+++ NEWS.md
@@ -1,3 +1,7 @@
+# PAnEn 3.5.1
+
+- Add the beta function `RAnEn::schaakeShuffle`.
+
 # PAnEn 3.5.0
 
 - Add functions `toDateTime`, `toForecastTime`, `val2Col`, `makeTransparent`, and `plotBackground`.